feat: Support ghc-9.10 ...
... assuming the existence of a 9.10-compatible version of
openapi3 < 3.3.

A ghc-9.10 version of openapi3 doesn't yet exist, so you won't be able to
get a build plan without adding some `allow-newer` to your cabal.project
file. But that's ok. Either a build-compatible version of openapi3 will
materialize at some point, or else the dependencies for this patch version
will just never be resolvable in the wild, which is not technically a
violation of PVP, so we're good.